diffraction: extract orientation histograms and flowlines
Third extraction. The cleanest cluster yet: of the 1564 lines, 678 referenced
`self` not once -- the three flowline renderers and both intensity helpers were
methods purely by accident. Nothing in it is polymer-specific; orientation
histograms and flowlines apply to any material with anisotropic scattering.

New `orientation.py` holds the pipeline -- make_orientation_histogram (threaded
polar_peaks + peak_intensities), make_flowline_map, and the rainbow/combined/
legend renderers -- plus `_get_intensity`/`_set_intensity`, now private module
helpers rather than public methods, which removes make_flowline_map's only tie to
the class. `plot_orientation_correlation` went to `orientation_correlation.py`
beside `calculate_orientation_correlation`, taking SLOPE_WEIGHT_FRACTION with it.

All eight public names keep forwarding methods.
`plot_orientation_correlation`'s wrapper preserves its self-fallback, so
`bp.plot_orientation_correlation()` with no arguments still plots the volume the
object last computed. `calculate_orientation_correlation` stays put: it writes
orient_corr/orient_corr_pairs onto the object, and already delegated its maths.

Verified per the protocol in dp_peak_detection's HANDOFF:
- Baseline tests written FIRST (e80ffa7) for the 678 previously uncovered lines.
- AST scan for rebinding of injected parameter names: none.
- refactor_diff_check.py: 23/23 bit-identical against baseline 1265551, the seven
  new orientation checks included.

One bug caught by the baseline tests: the generated wrapper prefixed `_` onto the
already-underscored `_set_intensity`, producing `__set_intensity`, which Python
name-mangles inside a class body. Fixed. Exactly the class of failure the
tests-first ordering is meant to catch, and it would have been invisible before
e80ffa7 since these had no coverage.

BraggPeaksPolymer: 4812 -> 3296 lines. bragg_peaks.py 6162 -> 4648, from 8206 at
the start of the refactor.

Tests: 500 passed, 10 skipped.
