+++ src/quantem/diffraction/bragg_peaks.py
@@ -52,8 +52,18 @@ from quantem.diffraction.vector_fields import (
     vector_field_cell as _vector_field_cell,
     vector_field_flat as _vector_field_flat,
 )
+from quantem.diffraction.orientation import (
+    _get_intensity,
+    _set_intensity,
+    make_flowline_combined_image as _make_flowline_combined_image,
+    make_flowline_map as _make_flowline_map,
+    make_flowline_rainbow_image as _make_flowline_rainbow_image,
+    make_flowline_rainbow_legend as _make_flowline_rainbow_legend,
+    make_orientation_histogram as _make_orientation_histogram,
+)
 from quantem.diffraction.orientation_correlation import (
     calculate_orientation_correlation as _calculate_orientation_correlation,
+    plot_orientation_correlation as _plot_orientation_correlation,
 )
 from quantem.core.utils.utils import electron_wavelength_angstrom
 from quantem.diffraction.polymer_utils import parse_reciprocal_units, sample_average_from_image
@@ -67,14 +77,6 @@ from pathlib import Path
 from mpl_toolkits.axes_grid1.inset_locator import inset_axes
 from matplotlib.patches import Ellipse, Rectangle
 from matplotlib.colors import BoundaryNorm, hsv_to_rgb, rgb_to_hsv
-
-# Default 1/e decay length for the orientation-correlation slope fit, as a
-# fraction of the fitted lobe's distance span. Chosen on real scan data
-# (pg3T2 07-03-2024/61): it holds the fitted intercept within ~1.7 degrees of
-# the measured boundary at zero separation across all six ring pairs, versus
-# up to 7 degrees for an unweighted fit, while keeping enough effective points
-# for a stable slope. See plot_orientation_correlation(slope_weight_scale=...).
-SLOPE_WEIGHT_FRACTION = 0.10
 
 
 def _apply_zoom_crop(data, zoom_factor, center=None):
@@ -3260,358 +3262,13 @@ class BraggPeaksPolymer(AutoSerialize):
 
         return _peak_radial_count_plot(self.polar_peaks, *args, **kwargs)
 
-    def make_orientation_histogram(
-        self,
-        radial_ranges: np.ndarray = None,
-        orientation_map=None,
-        orientation_ind: int = 0,
-        orientation_growth_angles: np.array = 0.0,
-        orientation_separate_bins: bool = False,
-        orientation_flip_sign: bool = False,
-        orientation_offset_degrees: float = 0.0,
-        upsample_factor: float = 4.0,
-        theta_step_deg: float = 1.0,
-        sigma_x: float = 1.0,
-        sigma_y: float = 1.0,
-        sigma_theta: float = 3.0,
-        use_peak_sigma: bool = False,
-        peak_sigma_samples: int = 6,
-        normalize_intensity_image: bool = False,
-        normalize_intensity_stack: bool = True,
-        progress_bar: bool = True,
-        r_field: str = "r_invA",
-        theta_field: str = "theta",
-        intensity_field: str = "intensities",
-        # intensity_field: str = "intensities_sampled_from_dp",
-    ):
+    def make_orientation_histogram(self, *args, **kwargs):
+        """Orientation histogram from this analysis's polar peaks.
+
+        Forwards to :func:`quantem.diffraction.orientation.make_orientation_histogram`.
         """
-        Create a 3D or 4D orientation histogram from bragg peaks.
-        
-        Can generate histograms from either:
-        1. Polar peak data with radial ranges
-        2. Orientation map with Euler angles (for fiber textures)
-    
-        Parameters
-        ----------
-        radial_ranges : np.ndarray, optional
-            Size (N x 2) array for N radial bins, or (2,) for a single bin.
-        orientation_map : OrientationMap, optional
-            Class containing Euler angles to generate a flowline map.
-        orientation_ind : int
-            Index of the orientation map (default 0)
-        orientation_growth_angles : np.array
-            Angles to place into histogram, relative to orientation.
-        orientation_separate_bins : bool
-            Whether to place multiple angles into multiple radial bins.
-        orientation_flip_sign : bool
-            Flip the direction of theta
-        orientation_offset_degrees : float
-            Offset for orientation angles in degrees
-        upsample_factor : float
-            Upsample factor for output histogram
-        theta_step_deg : float
-            Step size along annular direction in degrees
-        sigma_x : float
-            Smoothing in x direction before upsample
-        sigma_y : float
-            Smoothing in y direction before upsample
-        sigma_theta : float
-            Smoothing in annular direction (units of bins, periodic)
-        use_peak_sigma : bool
-            Spread signal along annular direction using measured peak width
-        peak_sigma_samples : int
-            Number of samples for peak sigma spreading
-        normalize_intensity_image : bool
-            Normalize to max peak intensity = 1, per image
-        normalize_intensity_stack : bool
-            Normalize to max peak intensity = 1, all images
-        progress_bar : bool
-            Enable progress bar
-        r_field : str
-            Name of radial coordinate field
-        theta_field : str
-            Name of angular coordinate field
-        intensity_field : str
-            Name of intensity field
-    
-        Returns
-        -------
-        orient_hist : np.ndarray
-            4D array containing Bragg peak intensity histogram
-            [radial_bin, x_probe, y_probe, theta]
-        """
-        # Coordinates
-        theta = np.arange(0, 180, theta_step_deg) * np.pi / 180.0
-        dtheta = theta[1] - theta[0]
-        dtheta_deg = dtheta * 180 / np.pi
-        num_theta_bins = np.size(theta)
-        
-        # Setup for peak sigma spreading
-        if use_peak_sigma:
-            v_sigma = np.linspace(-2, 2, 2 * peak_sigma_samples + 1)
-            w_sigma = np.exp(-(v_sigma**2) / 2)
-    
-        if orientation_map is None:
-            # Input bins
-            radial_ranges = np.array(radial_ranges)
-            if radial_ranges.ndim == 1:
-                radial_ranges = radial_ranges[None, :]
-            radial_ranges_2 = radial_ranges**2
-            num_radii = radial_ranges.shape[0]
-            size_input = self.polar_peaks.shape
-        else:
-            orientation_growth_angles = np.atleast_1d(orientation_growth_angles)
-            num_angles = orientation_growth_angles.shape[0]
-            size_input = [orientation_map.num_x, orientation_map.num_y]
-            if orientation_separate_bins is False:
-                num_radii = 1
-            else:
-                num_radii = num_angles
-    
-        size_output = np.round(
-            np.array(size_input).astype("float") * upsample_factor
-        ).astype("int")
-    
-        # Output init
-        orient_hist = np.zeros([num_radii, size_output[0], size_output[1], num_theta_bins])
-    
-        # Loop over all probe positions
-        for a0 in range(num_radii):
-            t = "Generating histogram " + str(a0)
-            for rx, ry in tqdmnd(
-                *size_input, desc=t, unit=" probe positions", disable=not progress_bar
-            ):
-                x = (rx + 0.5) * upsample_factor - 0.5
-                y = (ry + 0.5) * upsample_factor - 0.5
-                x = np.clip(x, 0, size_output[0] - 2)
-                y = np.clip(y, 0, size_output[1] - 2)
-                xF = np.floor(x).astype("int")
-                yF = np.floor(y).astype("int")
-                dx = x - xF
-                dy = y - yF
-    
-                add_data = False
-    
-                if orientation_map is None:
-                    p_r = _vector_field_cell(self.polar_peaks, r_field, rx, ry)
-                    p_theta = _vector_field_cell(self.polar_peaks, theta_field, rx, ry)
-                    
-                    if p_r is not None and len(p_r) > 0:
-                        r2 = p_r**2
-                        sub = np.logical_and(
-                            r2 >= radial_ranges_2[a0, 0], 
-                            r2 < radial_ranges_2[a0, 1]
-                        )
-                        if np.any(sub):
-                            intensity_data = _vector_field_cell(
-                                self.peak_intensities, intensity_field, rx, ry
-                            )
-                            if intensity_data is not None and len(intensity_data) > 0:
-                                add_data = True
-                                intensity = intensity_data[sub]
-                                
-                                # Get theta values
-                                theta_radians = p_theta[sub]
-                                if orientation_flip_sign:
-                                    theta_radians *= -1
-                                # Add offset
-                                theta_radians += orientation_offset_degrees * np.pi / 180
-                                theta_radians = np.mod(theta_radians, np.pi)
-                                t = theta_radians / dtheta
-                                
-                                # Spread signal using peak sigma if requested
-                                if use_peak_sigma:
-                                    # Try to get sigma values if available
-                                    if 'sigma_theta' in self.polar_peaks.fields:
-                                        theta_std = _vector_field_cell(
-                                            self.polar_peaks, "sigma_theta", rx, ry
-                                        )[sub] / dtheta
-                                        t = (t[:, None] + theta_std[:, None] * v_sigma[None, :]).ravel()
-                                        intensity = (intensity[:, None] * w_sigma[None, :]).ravel()
-                else:
-                    if orientation_map.corr[rx, ry, orientation_ind] > 0:
-                        if orientation_separate_bins is False:
-                            if orientation_flip_sign:
-                                t = (
-                                    np.array(
-                                        [
-                                            (
-                                                -orientation_map.angles[
-                                                    rx, ry, orientation_ind, 0
-                                                ]
-                                                - orientation_map.angles[
-                                                    rx, ry, orientation_ind, 2
-                                                ]
-                                            )
-                                            / dtheta
-                                        ]
-                                    )
-                                    + orientation_growth_angles
-                                )
-                            else:
-                                t = (
-                                    np.array(
-                                        [
-                                            (
-                                                orientation_map.angles[
-                                                    rx, ry, orientation_ind, 0
-                                                ]
-                                                + orientation_map.angles[
-                                                    rx, ry, orientation_ind, 2
-                                                ]
-                                            )
-                                            / dtheta
-                                        ]
-                                    )
-                                    + orientation_growth_angles
-                                )
-                            # Add offset
-                            t += orientation_offset_degrees / dtheta_deg
-                            intensity = (
-                                np.ones(num_angles)
-                                * orientation_map.corr[rx, ry, orientation_ind]
-                            )
-                            add_data = True
-                        else:
-                            if orientation_flip_sign:
-                                t = (
-                                    np.array(
-                                        [
-                                            (
-                                                -orientation_map.angles[
-                                                    rx, ry, orientation_ind, 0
-                                                ]
-                                                - orientation_map.angles[
-                                                    rx, ry, orientation_ind, 2
-                                                ]
-                                            )
-                                            / dtheta
-                                        ]
-                                    )
-                                    + orientation_growth_angles[a0]
-                                )
-                            else:
-                                t = (
-                                    np.array(
-                                        [
-                                            (
-                                                orientation_map.angles[
-                                                    rx, ry, orientation_ind, 0
-                                                ]
-                                                + orientation_map.angles[
-                                                    rx, ry, orientation_ind, 2
-                                                ]
-                                            )
-                                            / dtheta
-                                        ]
-                                    )
-                                    + orientation_growth_angles[a0]
-                                )
-                            # Add offset
-                            t += orientation_offset_degrees / dtheta_deg
-                            intensity = orientation_map.corr[rx, ry, orientation_ind]
-                            add_data = True
-    
-                if add_data:
-                    tF = np.floor(t).astype("int")
-                    dt = t - tF
-    
-                    orient_hist[a0, xF, yF, :] = orient_hist[a0, xF, yF, :] + np.bincount(
-                        np.mod(tF, num_theta_bins),
-                        weights=(1 - dx) * (1 - dy) * (1 - dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-                    orient_hist[a0, xF, yF, :] = orient_hist[a0, xF, yF, :] + np.bincount(
-                        np.mod(tF + 1, num_theta_bins),
-                        weights=(1 - dx) * (1 - dy) * (dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-    
-                    orient_hist[a0, xF + 1, yF, :] = orient_hist[
-                        a0, xF + 1, yF, :
-                    ] + np.bincount(
-                        np.mod(tF, num_theta_bins),
-                        weights=(dx) * (1 - dy) * (1 - dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-                    orient_hist[a0, xF + 1, yF, :] = orient_hist[
-                        a0, xF + 1, yF, :
-                    ] + np.bincount(
-                        np.mod(tF + 1, num_theta_bins),
-                        weights=(dx) * (1 - dy) * (dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-    
-                    orient_hist[a0, xF, yF + 1, :] = orient_hist[
-                        a0, xF, yF + 1, :
-                    ] + np.bincount(
-                        np.mod(tF, num_theta_bins),
-                        weights=(1 - dx) * (dy) * (1 - dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-                    orient_hist[a0, xF, yF + 1, :] = orient_hist[
-                        a0, xF, yF + 1, :
-                    ] + np.bincount(
-                        np.mod(tF + 1, num_theta_bins),
-                        weights=(1 - dx) * (dy) * (dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-    
-                    orient_hist[a0, xF + 1, yF + 1, :] = orient_hist[
-                        a0, xF + 1, yF + 1, :
-                    ] + np.bincount(
-                        np.mod(tF, num_theta_bins),
-                        weights=(dx) * (dy) * (1 - dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-                    orient_hist[a0, xF + 1, yF + 1, :] = orient_hist[
-                        a0, xF + 1, yF + 1, :
-                    ] + np.bincount(
-                        np.mod(tF + 1, num_theta_bins),
-                        weights=(dx) * (dy) * (dt) * intensity,
-                        minlength=num_theta_bins,
-                    )
-    
-        # Smoothing / interpolation
-        if (sigma_x is not None) or (sigma_y is not None) or (sigma_theta is not None):
-            if num_radii > 1:
-                print("Interpolating orientation matrices ...", end="")
-            else:
-                print("Interpolating orientation matrix ...", end="")
-            if sigma_x is not None and sigma_x > 0:
-                orient_hist = gaussian_filter1d(
-                    orient_hist,
-                    sigma_x * upsample_factor,
-                    mode="nearest",
-                    axis=1,
-                    truncate=3.0,
-                )
-            if sigma_y is not None and sigma_y > 0:
-                orient_hist = gaussian_filter1d(
-                    orient_hist,
-                    sigma_y * upsample_factor,
-                    mode="nearest",
-                    axis=2,
-                    truncate=3.0,
-                )
-            if sigma_theta is not None and sigma_theta > 0:
-                orient_hist = gaussian_filter1d(
-                    orient_hist, sigma_theta / dtheta_deg, mode="wrap", axis=3, truncate=2.0
-                )
-            print(" done.")
-    
-        # Normalization
-        if normalize_intensity_stack is True:
-            stack_max = np.max(orient_hist)
-            if stack_max > 0:
-                orient_hist = orient_hist / stack_max
-        elif normalize_intensity_image is True:
-            for a0 in range(num_radii):
-                image_max = np.max(orient_hist[a0, :, :, :])
-                if image_max > 0:
-                    orient_hist[a0, :, :, :] /= image_max
-    
-        return orient_hist
+
+        return _make_orientation_histogram(self.polar_peaks, self.peak_intensities, *args, **kwargs)
 
     def calculate_orientation_correlation(
         self,
@@ -3707,450 +3364,20 @@ class BraggPeaksPolymer(AutoSerialize):
             self.orient_corr_pairs = pair_indices
         return orient_corr
 
-    def plot_orientation_correlation(
-        self,
-        orient_corr=None,
-        *,
-        pair_indices=None,
-        pixel_size=1.0,
-        pixel_units="scan pixels",
-        probability_range=(0.5, 2.0),
-        cmap="correlation",
-        figsize=None,
-        show_metrics=True,
-        return_metrics=False,
-        slope_weight_scale=None,
-    ):
-        """Plot distance-orientation correlations using Matplotlib.
+    def plot_orientation_correlation(self, orient_corr=None, *args, **kwargs):
+        """Plot the distance-orientation correlation.
 
-        The 50% boundary is halfway between the correlation at zero separation
-        and the random-association baseline of one. Its intercepts give the
-        radial and annular 50% distances. The signed slope is fitted separately
-        to the primary correlation-equals-one boundary between positive
-        correlation and anticorrelation.
-
-        Parameters
-        ----------
-        slope_weight_scale : float, optional
-            1/e decay length, in ``pixel_units``, of the exponential weighting
-            applied to the slope fit. The correlation-equals-one boundary
-            saturates with distance, so an unweighted straight line over the
-            whole lobe is dominated by the flat tail: it biases the slope low
-            and pushes the fitted intercept off the measured boundary at zero
-            separation. Weighting towards short distances makes ``slope`` the
-            near-origin tangent instead. Defaults to
-            ``SLOPE_WEIGHT_FRACTION`` times the fitted distance span. Pass
-            ``numpy.inf`` to restore the previous unweighted full-lobe fit.
-
-        Notes
-        -----
-        Each metrics entry reports ``slope_fit_intercept_degrees`` (compare it
-        against the boundary at zero separation to check the fit),
-        ``slope_fit_effective_point_count`` (Kish effective sample size, which
-        falls as the weighting sharpens), and the resolved
-        ``slope_weight_scale``. ``slope_fit_r_squared`` is weighted with the
-        same weights, so it is not comparable to an unweighted R-squared.
+        Forwards to
+        :func:`quantem.diffraction.orientation_correlation.plot_orientation_correlation`,
+        defaulting to the correlation volume this object last computed.
         """
-        from matplotlib.colors import LinearSegmentedColormap, LogNorm
-        from matplotlib.lines import Line2D
 
-        def crossing(coordinates, profile, level):
-            profile = np.asarray(profile, dtype=float)
-            coordinates = np.asarray(coordinates, dtype=float)
-            if not np.isfinite(profile[0]):
-                return np.nan
-            initial_side = profile[0] - level
-            if initial_side == 0:
-                return float(coordinates[0])
-            for point in range(1, len(profile)):
-                before, after = profile[point - 1], profile[point]
-                if not np.isfinite(before) or not np.isfinite(after):
-                    continue
-                before_side = before - level
-                after_side = after - level
-                if before_side == 0:
-                    return float(coordinates[point - 1])
-                if before_side * after_side <= 0:
-                    if before == after:
-                        return float(coordinates[point])
-                    fraction = -before_side / (after_side - before_side)
-                    return float(
-                        coordinates[point - 1]
-                        + fraction * (coordinates[point] - coordinates[point - 1])
-                    )
-            return np.nan
-
-        values = self.orient_corr if orient_corr is None else orient_corr
-        if values is None:
-            raise RuntimeError(
-                "No orientation correlation is available. Run "
-                "calculate_orientation_correlation() first or pass orient_corr."
-            )
-        values = np.asarray(values)
-        if values.ndim != 3:
-            raise ValueError(
-                "orient_corr must have shape (pair, relative_angle, distance)."
-            )
-        if values.shape[0] == 0:
-            raise ValueError("orient_corr must contain at least one radial-bin pair.")
-        if values.shape[1] < 2 or values.shape[2] < 2:
-            raise ValueError(
-                "orient_corr requires at least two angle and two distance samples."
-            )
-        labels = self.orient_corr_pairs if pair_indices is None else pair_indices
-        if labels is not None:
-            labels = np.asarray(labels)
-            if labels.shape != (values.shape[0], 2):
-                raise ValueError(
-                    f"pair_indices must have shape ({values.shape[0]}, 2)."
-                )
-
-        panel_count = values.shape[0]
-        column_count = min(3, max(1, panel_count))
-        row_count = int(np.ceil(panel_count / column_count))
-        if figsize is None:
-            figsize = (4.5 * column_count, 3.8 * row_count)
-        fig, axes = plt.subplots(
-            row_count,
-            column_count,
-            figsize=figsize,
-            squeeze=False,
-            constrained_layout=True,
+        return _plot_orientation_correlation(
+            self.orient_corr if orient_corr is None else orient_corr,
+            self.orient_corr_pairs,
+            *args,
+            **kwargs,
         )
-        lower, upper = map(float, probability_range)
-        if not 0 < lower < upper:
-            raise ValueError("probability_range must satisfy 0 < lower < upper.")
-        if cmap == "correlation":
-            cmap = LinearSegmentedColormap.from_list(
-                "quantem_correlation",
-                [
-                    (0.00, "#002b9a"),
-                    (0.32, "#1769e8"),
-                    (0.50, "#b8b8b8"),
-                    (0.68, "#f23838"),
-                    (1.00, "#9e0015"),
-                ],
-            )
-        distance_max = (values.shape[2] - 1) * float(pixel_size)
-        distances = np.arange(values.shape[2], dtype=float) * float(pixel_size)
-        angles = np.linspace(0.0, 180.0, values.shape[1])
-        image = None
-        metrics = []
-        for index, ax in enumerate(axes.flat):
-            if index >= panel_count:
-                ax.set_visible(False)
-                continue
-            image = ax.imshow(
-                values[index],
-                origin="lower",
-                aspect="auto",
-                extent=(0, distance_max, 0, 180),
-                norm=LogNorm(vmin=lower, vmax=upper),
-                cmap=cmap,
-            )
-            if labels is None:
-                title = f"Ring pair {index}"
-                pair = (index, index)
-            else:
-                pair = tuple(int(value) for value in labels[index])
-                title = (
-                    f"Autocorrelation of Ring {pair[0]}"
-                    if pair[0] == pair[1]
-                    else f"Correlation of Rings {pair[0]} and {pair[1]}"
-                )
-            ax.set(
-                title=title,
-                xlabel=f"distance ({pixel_units})",
-                ylabel="relative orientation (degrees)",
-            )
-            panel = np.asarray(values[index], dtype=float)
-            origin_probability = panel[0, 0]
-            half_probability = (
-                1.0 + 0.5 * (origin_probability - 1.0)
-                if np.isfinite(origin_probability)
-                and origin_probability > 0.0
-                and not np.isclose(origin_probability, 1.0)
-                else np.nan
-            )
-            radial_distance = np.nan
-            annular_distance = np.nan
-            slope = np.nan
-            slope_fit_r_squared = np.nan
-            slope_fit_point_count = 0
-            slope_fit_intercept = np.nan
-            slope_fit_effective_count = np.nan
-            weight_scale = np.nan
-            fit_distances = np.array([])
-            fit_angles = np.array([])
-            if np.isfinite(half_probability):
-                radial_distance = crossing(
-                    distances, panel[0, :], half_probability
-                )
-                annular_distance = crossing(
-                    angles, panel[:, 0], half_probability
-                )
-                if np.isfinite(radial_distance):
-                    ax.scatter(
-                        [radial_distance],
-                        [0],
-                        marker="o",
-                        s=45,
-                        facecolor="white",
-                        edgecolor="black",
-                        linewidth=0.8,
-                        zorder=5,
-                    )
-                if np.isfinite(annular_distance):
-                    ax.scatter(
-                        [0],
-                        [annular_distance],
-                        marker="D",
-                        s=40,
-                        facecolor="white",
-                        edgecolor="black",
-                        linewidth=0.8,
-                        zorder=5,
-                    )
-
-            # The slope belongs to the gray probability/random = 1 boundary, not
-            # the half-maximum contour used for the two distance intercepts.
-            baseline_boundary = np.array(
-                [
-                    crossing(angles, panel[:, radius], 1.0)
-                    for radius in range(panel.shape[1])
-                ]
-            )
-            baseline_radial_intercept = crossing(distances, panel[0, :], 1.0)
-            valid = np.isfinite(baseline_boundary)
-            if np.isfinite(baseline_radial_intercept):
-                valid &= distances <= baseline_radial_intercept + float(pixel_size)
-
-            # Select the earliest contiguous run: this follows the principal
-            # red/blue lobe from the angular axis and rejects remote closed loops.
-            valid_indices = np.flatnonzero(valid)
-            primary_indices = np.array([], dtype=int)
-            if valid_indices.size:
-                angular_jump = np.abs(
-                    np.diff(baseline_boundary[valid_indices])
-                )
-                maximum_step = max(15.0, 4.0 * (angles[1] - angles[0]))
-                split_points = np.flatnonzero(
-                    (np.diff(valid_indices) > 1) | (angular_jump > maximum_step)
-                ) + 1
-                segments = np.split(
-                    valid_indices, split_points
-                )
-                primary_indices = next(
-                    (segment for segment in segments if len(segment) >= 2),
-                    np.array([], dtype=int),
-                )
-            if primary_indices.size >= 5:
-                # A connected correlation=1 contour can rise away from the
-                # origin, turn around at large distance, and return as part of
-                # the same loop. Fitting that entire loop can reverse the sign
-                # of the visually obvious near-origin boundary. Stop at the
-                # first sustained turning point while tolerating isolated
-                # pixel-scale contour noise.
-                boundary_run = baseline_boundary[primary_indices]
-                smoothing_sigma = min(3.0, max(0.75, len(boundary_run) / 100.0))
-                boundary_smooth = gaussian_filter1d(
-                    boundary_run, sigma=smoothing_sigma, mode="nearest"
-                )
-                boundary_gradient = np.gradient(boundary_smooth)
-                initial_count = min(20, max(3, len(boundary_run) // 10))
-                initial_trend = float(
-                    np.median(boundary_gradient[:initial_count])
-                )
-                if not np.isclose(initial_trend, 0.0):
-                    reversal = boundary_gradient * np.sign(initial_trend) < 0
-                    persistence = min(5, max(2, len(boundary_run) // 20))
-                    sustained = np.convolve(
-                        reversal.astype(int),
-                        np.ones(persistence, dtype=int),
-                        mode="valid",
-                    )
-                    turning_points = np.flatnonzero(sustained == persistence)
-                    if turning_points.size and turning_points[0] >= 2:
-                        primary_indices = primary_indices[
-                            : turning_points[0] + 1
-                        ]
-            if primary_indices.size:
-                fit_distances = distances[primary_indices]
-                fit_values = baseline_boundary[primary_indices]
-
-                # The correlation=1 boundary saturates: it climbs steeply near the
-                # origin and flattens at large separation. An unweighted straight
-                # line over the whole lobe is therefore dominated by the flat tail,
-                # which drags the intercept off the measured boundary at d=0 (up to
-                # ~7 degrees on real data, visibly landing in the blue region) and
-                # biases the slope low. Weight the fit towards short distances so
-                # `slope` is the near-origin tangent, which is the physically
-                # meaningful quantity.
-                fit_span = float(fit_distances.max() - fit_distances.min())
-                if slope_weight_scale is None:
-                    weight_scale = SLOPE_WEIGHT_FRACTION * fit_span
-                else:
-                    weight_scale = float(slope_weight_scale)
-                if not np.isfinite(weight_scale) or weight_scale <= 0:
-                    # np.inf (or a non-positive scale) restores the legacy
-                    # unweighted fit over the full lobe.
-                    fit_weights = np.ones_like(fit_distances)
-                    weight_scale = np.inf
-                else:
-                    fit_weights = np.exp(
-                        -(fit_distances - fit_distances.min()) / weight_scale
-                    )
-
-                root_weights = np.sqrt(fit_weights)
-                design = (
-                    np.vstack([fit_distances, np.ones_like(fit_distances)]).T
-                    * root_weights[:, None]
-                )
-                fit_slope, fit_intercept = np.linalg.lstsq(
-                    design, fit_values * root_weights, rcond=None
-                )[0]
-                fit_angles = fit_intercept + fit_slope * fit_distances
-                slope = float(fit_slope)
-                slope_fit_intercept = float(fit_intercept)
-                slope_fit_point_count = int(fit_distances.size)
-                # Kish effective sample size: how many points the weighting really
-                # uses, so a too-aggressive scale is visible rather than silent.
-                slope_fit_effective_count = float(
-                    fit_weights.sum() ** 2 / np.sum(fit_weights**2)
-                )
-                weight_mean = float(
-                    np.average(fit_values, weights=fit_weights)
-                )
-                residual_sum_squares = float(
-                    np.sum(fit_weights * (fit_values - fit_angles) ** 2)
-                )
-                total_sum_squares = float(
-                    np.sum(fit_weights * (fit_values - weight_mean) ** 2)
-                )
-                slope_fit_r_squared = (
-                    1.0 - residual_sum_squares / total_sum_squares
-                    if total_sum_squares > 0
-                    else np.nan
-                )
-
-            if fit_distances.size:
-                # Draw only where the weighting actually constrains the line. Past
-                # ~3 decay lengths the boundary has saturated away from this
-                # tangent, and extending the line there would misrepresent the fit.
-                if np.isfinite(weight_scale):
-                    drawn = fit_distances <= (
-                        fit_distances.min() + 3.0 * weight_scale
-                    )
-                    if drawn.sum() < 2:
-                        drawn = np.zeros_like(fit_distances, dtype=bool)
-                        drawn[: min(2, drawn.size)] = True
-                else:
-                    drawn = np.ones_like(fit_distances, dtype=bool)
-                visible_fit = drawn & (fit_angles >= 0) & (fit_angles <= 180)
-                ax.plot(
-                    fit_distances[visible_fit],
-                    fit_angles[visible_fit],
-                    color="#ffe600",
-                    linestyle="-",
-                    linewidth=2.5,
-                    zorder=6,
-                )
-            ax.legend(
-                handles=[
-                    Line2D(
-                        [0],
-                        [0],
-                        marker="o",
-                        color="none",
-                        markerfacecolor="white",
-                        markeredgecolor="black",
-                        label="50% radial intercept",
-                    ),
-                    Line2D(
-                        [0],
-                        [0],
-                        marker="D",
-                        color="none",
-                        markerfacecolor="white",
-                        markeredgecolor="black",
-                        label="50% annular intercept",
-                    ),
-                    Line2D(
-                        [0],
-                        [0],
-                        color="#ffe600",
-                        linewidth=2.5,
-                        label="near-origin baseline fit",
-                    ),
-                ],
-                loc="lower right",
-                fontsize=7,
-                framealpha=0.82,
-            )
-
-            panel_metrics = {
-                "pair": pair,
-                "title": title,
-                "half_probability": float(half_probability),
-                "radial_distance": float(radial_distance),
-                "annular_distance_degrees": float(annular_distance),
-                "slope_degrees_per_unit": float(slope),
-                "slope_fit_r_squared": float(slope_fit_r_squared),
-                "slope_fit_point_count": slope_fit_point_count,
-                "slope_fit_intercept_degrees": float(slope_fit_intercept),
-                "slope_fit_effective_point_count": float(
-                    slope_fit_effective_count
-                ),
-                "slope_weight_scale": float(weight_scale),
-                "slope_contour_probability": 1.0,
-                "distance_units": pixel_units,
-            }
-            metrics.append(panel_metrics)
-            if show_metrics:
-                radial_text = (
-                    f"{radial_distance:.2f} {pixel_units}"
-                    if np.isfinite(radial_distance)
-                    else "not resolved"
-                )
-                annular_text = (
-                    f"{annular_distance:.2f} degrees"
-                    if np.isfinite(annular_distance)
-                    else "not resolved"
-                )
-                slope_text = (
-                    f"{slope:.2f} degrees/{pixel_units}"
-                    if np.isfinite(slope)
-                    else "not resolved"
-                )
-                ax.text(
-                    0.98,
-                    0.98,
-                    "50% radial distance = "
-                    + radial_text
-                    + "\n50% annular distance = "
-                    + annular_text
-                    + "\nslope = "
-                    + slope_text,
-                    transform=ax.transAxes,
-                    ha="right",
-                    va="top",
-                    fontsize=8,
-                    bbox={
-                        "boxstyle": "round,pad=0.35",
-                        "facecolor": "white",
-                        "edgecolor": "black",
-                        "alpha": 0.82,
-                    },
-                )
-        if image is not None:
-            fig.colorbar(
-                image,
-                ax=[ax for ax in axes.flat if ax.get_visible()],
-                label="probability / random",
-            )
-        if return_metrics:
-            return fig, axes, metrics
-        return fig, axes
 
     def plot_interactive_image_map(self, *args, **kwargs):
         """Interactive diffraction-pattern browser.
@@ -5088,794 +4315,53 @@ class BraggPeaksPolymer(AutoSerialize):
 
         return _plot_peak_count_map(self.peak_coordinates_cartesian, self.polar_peaks, *args, **kwargs)
 
-    def make_flowline_map(
-        self,
-        orient_hist,
-        thresh_seed=0.2,
-        thresh_grow=0.05,
-        thresh_collision=0.001,
-        sep_seeds=None,
-        sep_xy=6.0,
-        sep_theta=5.0,
-        sort_seeds="intensity",
-        linewidth=2.0,
-        step_size=0.5,
-        min_steps=4,
-        max_steps=1000,
-        sigma_x=1.0,
-        sigma_y=1.0,
-        sigma_theta=2.0,
-        progress_bar: bool = True,
-    ):
-        """
-        Create an 3D or 4D orientation flowline map - essentially a pixelated "stream map" which represents diffraction data.
-    
-        Args:
-            orient_hist (array):        Histogram of all orientations with coordinates
-                                        [radial_bin x_probe y_probe theta]
-                                        We assume theta bin ranges from 0 to 180 degrees and is periodic.
-            thresh_seed (float):        Threshold for seed generation in histogram.
-            thresh_grow (float):        Threshold for flowline growth in histogram.
-            thresh_collision (float):   Threshold for termination of flowline growth in histogram.
-            sep_seeds (float):          Initial seed separation in bins - set to None to use default value,
-                                        which is equal to 0.5*sep_xy.
-            sep_xy (float):             Search radius for flowline direction in x and y.
-            sep_theta = (float):        Search radius for flowline direction in theta.
-            sort_seeds (str):           How to sort the initial seeds for growth:
-                                            None - no sorting
-                                            'intensity' - sort by histogram intensity
-                                            'random' - random order
-            linewidth (float):          Thickness of the flowlines in pixels.
-            step_size (float):          Step size for flowline growth in pixels.
-            min_steps (int):            Minimum number of steps for a flowline to be drawn.
-            max_steps (int):            Maximum number of steps for a flowline to be drawn.
-            sigma_x (float):            Weighted sigma in x direction for direction update.
-            sigma_y (float):            Weighted sigma in y direction for direction update.
-            sigma_theta (float):        Weighted sigma in theta for direction update.
-            progress_bar (bool):        Enable progress bar
-    
-        Returns:
-            orient_flowlines (array):   4D array containing flowlines
-                                        [radial_bin x_probe y_probe theta]
-        """
-    
-        # Ensure sep_xy and sep_theta are arrays
-        sep_xy = np.atleast_1d(sep_xy)
-        sep_theta = np.atleast_1d(sep_theta)
-    
-        # number of radial bins
-        num_radii = orient_hist.shape[0]
-        if num_radii > 1 and len(sep_xy) == 1:
-            sep_xy = np.ones(num_radii) * sep_xy
-        if num_radii > 1 and len(sep_theta) == 1:
-            sep_theta = np.ones(num_radii) * sep_theta
-    
-        # Default seed separation
-        if sep_seeds is None:
-            sep_seeds = np.round(np.min(sep_xy) / 2 + 0.5).astype("int")
-        else:
-            sep_seeds = np.atleast_1d(sep_seeds).astype("int")
-            if num_radii > 1 and len(sep_seeds) == 1:
-                sep_seeds = (np.ones(num_radii) * sep_seeds).astype("int")
-    
-        # coordinates
-        theta = np.linspace(0, np.pi, orient_hist.shape[3], endpoint=False)
-        dtheta = theta[1] - theta[0]
-        size_3D = np.array(
-            [
-                orient_hist.shape[1],
-                orient_hist.shape[2],
-                orient_hist.shape[3],
-            ]
-        )
-    
-        # initialize weighting array
-        vx = np.arange(-np.ceil(2 * sigma_x), np.ceil(2 * sigma_x) + 1)
-        vy = np.arange(-np.ceil(2 * sigma_y), np.ceil(2 * sigma_y) + 1)
-        vt = np.arange(-np.ceil(2 * sigma_theta), np.ceil(2 * sigma_theta) + 1)
-        ay, ax, at = np.meshgrid(vy, vx, vt)
-        k = (
-            np.exp(ax**2 / (-2 * sigma_x**2))
-            * np.exp(ay**2 / (-2 * sigma_y**2))
-            * np.exp(at**2 / (-2 * sigma_theta**2))
-        )
-        k = k / np.sum(k)
-        vx = vx[:, None, None].astype("int")
-        vy = vy[None, :, None].astype("int")
-        vt = vt[None, None, :].astype("int")
-    
-        # initalize flowline array
-        orient_flowlines = np.zeros_like(orient_hist)
-    
-        # initialize output
-        xy_t_int = np.zeros((max_steps + 1, 4))
-        xy_t_int_rev = np.zeros((max_steps + 1, 4))
-    
-        # Loop over radial bins
-        for a0 in range(num_radii):
-            # initialize collision check array
-            cr = np.arange(-np.ceil(sep_xy[a0]), np.ceil(sep_xy[a0]) + 1)
-            ct = np.arange(-np.ceil(sep_theta[a0]), np.ceil(sep_theta[a0]) + 1)
-            ay, ax, at = np.meshgrid(cr, cr, ct)
-            c_mask = (
-                (ax**2 + ay**2) / sep_xy[a0] ** 2 + at**2 / sep_theta[a0] ** 2
-                <= (1 + 1 / sep_xy[a0]) ** 2
-            )[None, :, :, :]
-            cx = cr[None, :, None, None].astype("int")
-            cy = cr[None, None, :, None].astype("int")
-            ct = ct[None, None, None, :].astype("int")
-    
-            # Find all seed locations
-            orient = orient_hist[a0, :, :, :]
-            sub_seeds = np.logical_and(
-                np.logical_and(
-                    orient >= np.roll(orient, 1, axis=2),
-                    orient >= np.roll(orient, -1, axis=2),
-                ),
-                orient >= thresh_seed,
-            )
-    
-            # Separate seeds
-            if sep_seeds > 0:
-                for a1 in range(sep_seeds - 1):
-                    sub_seeds[a1::sep_seeds, :, :] = False
-                    sub_seeds[:, a1::sep_seeds, :] = False
-    
-            # Index seeds
-            x_inds, y_inds, t_inds = np.where(sub_seeds)
-            if sort_seeds is not None:
-                if sort_seeds == "intensity":
-                    inds_sort = np.argsort(orient[sub_seeds])[::-1]
-                elif sort_seeds == "random":
-                    inds_sort = np.random.permutation(np.count_nonzero(sub_seeds))
-                x_inds = x_inds[inds_sort]
-                y_inds = y_inds[inds_sort]
-                t_inds = t_inds[inds_sort]
-    
-            # for a1 in tqdmnd(range(0,40), desc="Drawing flowlines",unit=" seeds", disable=not progress_bar):
-            t = "Drawing flowlines " + str(a0)
-            for a1 in tqdmnd(
-                range(0, x_inds.shape[0]), desc=t, unit=" seeds", disable=not progress_bar
-            ):
-                # initial coordinate and intensity
-                xy0 = np.array((x_inds[a1], y_inds[a1]))
-                t0 = theta[t_inds[a1]]
-    
-                # init theta
-                inds_theta = np.mod(
-                    np.round(t0 / dtheta).astype("int") + vt, orient.shape[2]
-                )
-                orient_crop = (
-                    k
-                    * orient[
-                        np.clip(
-                            np.round(xy0[0]).astype("int") + vx, 0, orient.shape[0] - 1
-                        ),
-                        np.clip(
-                            np.round(xy0[1]).astype("int") + vy, 0, orient.shape[1] - 1
-                        ),
-                        inds_theta,
-                    ]
-                )
-                theta_crop = theta[inds_theta]
-                t0 = np.sum(orient_crop * theta_crop) / np.sum(orient_crop)
-    
-                # forward direction
-                t = t0
-                v0 = np.array((np.cos(t), -np.sin(t)))
-                v = v0 * step_size
-                xy = xy0
-                int_val = self.get_intensity(orient, xy0[0], xy0[1], t0 / dtheta)
-                xy_t_int[0, 0:2] = xy0
-                xy_t_int[0, 2] = t / dtheta
-                xy_t_int[0, 3] = int_val
-                # main loop
-                grow = True
-                count = 0
-                while grow is True:
-                    count += 1
-    
-                    # update position and intensity
-                    xy = xy + v
-                    int_val = self.get_intensity(orient, xy[0], xy[1], t / dtheta)
-    
-                    # check for collision
-                    flow_crop = orient_flowlines[
-                        a0,
-                        np.clip(np.round(xy[0]).astype("int") + cx, 0, orient.shape[0] - 1),
-                        np.clip(np.round(xy[1]).astype("int") + cy, 0, orient.shape[1] - 1),
-                        np.mod(np.round(t / dtheta).astype("int") + ct, orient.shape[2]),
-                    ]
-                    int_flow = np.max(flow_crop[c_mask])
-    
-                    if (
-                        xy[0] < 0
-                        or xy[1] < 0
-                        or xy[0] > orient.shape[0]
-                        or xy[1] > orient.shape[1]
-                        or int_val < thresh_grow
-                        or int_flow > thresh_collision
-                    ):
-                        grow = False
-                    else:
-                        # update direction
-                        inds_theta = np.mod(
-                            np.round(t / dtheta).astype("int") + vt, orient.shape[2]
-                        )
-                        orient_crop = (
-                            k
-                            * orient[
-                                np.clip(
-                                    np.round(xy[0]).astype("int") + vx,
-                                    0,
-                                    orient.shape[0] - 1,
-                                ),
-                                np.clip(
-                                    np.round(xy[1]).astype("int") + vy,
-                                    0,
-                                    orient.shape[1] - 1,
-                                ),
-                                inds_theta,
-                            ]
-                        )
-                        theta_crop = theta[inds_theta]
-                        t = np.sum(orient_crop * theta_crop) / np.sum(orient_crop)
-                        # v = np.array((np.cos(t), np.sin(t))) * step_size
-                        # v = np.array((np.sin(t), np.cos(t))) * step_size
-                        # v = np.array((-np.sin(t), np.cos(t))) * step_size
-    
-                        xy_t_int[count, 0:2] = xy
-                        xy_t_int[count, 2] = t / dtheta
-                        xy_t_int[count, 3] = int_val
-    
-                        if count > max_steps - 1:
-                            grow = False
-    
-                # reverse direction
-                t = t0 + np.pi
-                v0 = np.array((np.cos(t), -np.sin(t)))
-                v = v0 * step_size
-                xy = xy0
-                int_val = self.get_intensity(orient, xy0[0], xy0[1], t0 / dtheta)
-                xy_t_int_rev[0, 0:2] = xy0
-                xy_t_int_rev[0, 2] = t / dtheta
-                xy_t_int_rev[0, 3] = int_val
-                # main loop
-                grow = True
-                count_rev = 0
-                while grow is True:
-                    count_rev += 1
-    
-                    # update position and intensity
-                    xy = xy + v
-                    int_val = self.get_intensity(orient, xy[0], xy[1], t / dtheta)
-    
-                    # check for collision
-                    flow_crop = orient_flowlines[
-                        a0,
-                        np.clip(np.round(xy[0]).astype("int") + cx, 0, orient.shape[0] - 1),
-                        np.clip(np.round(xy[1]).astype("int") + cy, 0, orient.shape[1] - 1),
-                        np.mod(np.round(t / dtheta).astype("int") + ct, orient.shape[2]),
-                    ]
-                    int_flow = np.max(flow_crop[c_mask])
-    
-                    if (
-                        xy[0] < 0
-                        or xy[1] < 0
-                        or xy[0] > orient.shape[0]
-                        or xy[1] > orient.shape[1]
-                        or int_val < thresh_grow
-                        or int_flow > thresh_collision
-                    ):
-                        grow = False
-                    else:
-                        # update direction
-                        inds_theta = np.mod(
-                            np.round(t / dtheta).astype("int") + vt, orient.shape[2]
-                        )
-                        orient_crop = (
-                            k
-                            * orient[
-                                np.clip(
-                                    np.round(xy[0]).astype("int") + vx,
-                                    0,
-                                    orient.shape[0] - 1,
-                                ),
-                                np.clip(
-                                    np.round(xy[1]).astype("int") + vy,
-                                    0,
-                                    orient.shape[1] - 1,
-                                ),
-                                inds_theta,
-                            ]
-                        )
-                        theta_crop = theta[inds_theta]
-                        t = np.sum(orient_crop * theta_crop) / np.sum(orient_crop) + np.pi
-                        v = np.array((np.cos(t), -np.sin(t))) * step_size
-                        v = np.array((np.cos(t), -np.sin(t))) * step_size
-                        # v = np.array((-np.sin(t), np.cos(t))) * step_size
-    
-                        xy_t_int_rev[count_rev, 0:2] = xy
-                        xy_t_int_rev[count_rev, 2] = t / dtheta
-                        xy_t_int_rev[count_rev, 3] = int_val
-    
-                        if count_rev > max_steps - 1:
-                            grow = False
-    
-                # write into output array
-                if count + count_rev > min_steps:
-                    if count > 0:
-                        orient_flowlines[a0, :, :, :] = self.set_intensity(
-                            orient_flowlines[a0, :, :, :], xy_t_int[1:count, :]
-                        )
-                    if count_rev > 1:
-                        orient_flowlines[a0, :, :, :] = self.set_intensity(
-                            orient_flowlines[a0, :, :, :], xy_t_int_rev[1:count_rev, :]
-                        )
-    
-        # normalize to step size
-        orient_flowlines = orient_flowlines * step_size
-    
-        # linewidth
-        if linewidth > 1.0:
-            s = linewidth - 1.0
-    
-            orient_flowlines = gaussian_filter1d(orient_flowlines, s, axis=1, truncate=3.0)
-            orient_flowlines = gaussian_filter1d(orient_flowlines, s, axis=2, truncate=3.0)
-            orient_flowlines = orient_flowlines * (s**2)
-    
-        return orient_flowlines
-    
-    
-    def make_flowline_rainbow_image(
-        self,
-        orient_flowlines,
-        int_range=[0, 0.2],
-        sym_rotation_order=2,
-        theta_offset=np.pi,
-        greyscale=False,
-        greyscale_max=True,
-        white_background=False,
-        power_scaling=1.0,
-        sum_radial_bins=False,
-        plot_images=True,
-        figsize=None,
-    ):
-        """
-        Generate RGB output images from the flowline arrays.
-    
-        Args:
-            orient_flowline (array):    Histogram of all orientations with coordinates [x y radial_bin theta]
-                                        We assume theta bin ranges from 0 to 180 degrees and is periodic.
-            int_range (float)           2 element array giving the intensity range
-            sym_rotation_order (int):   rotational symmety for colouring
-            theta_offset (float):       Offset the anglular coloring by this value in radians.
-                                        Default pi rotates the hue mapping by 90deg (nematic
-                                        sym=2) so color tracks the drawn flowline direction:
-                                        cyan for vertical lines, red for horizontal.
-            greyscale (bool):           Set to False for color output, True for greyscale output.
-            greyscale_max (bool):       If output is greyscale, use max instead of mean for overlapping flowlines.
-            white_background (bool):    For either color or greyscale output, switch to white background (from black).
-            power_scaling (float):      Power law scaling for flowline intensity output.
-            sum_radial_bins (bool):     Sum all radial bins (alternative is to output separate images).
-            plot_images (bool):         Plot the outputs for quick visualization.
-            figsize (2-tuple):          Size of output figure.
-    
-        Returns:
-            im_flowline (array):        3D or 4D array containing flowline images
-        """
-    
-        # init array
-        size_input = orient_flowlines.shape
-        size_output = np.array([size_input[0], size_input[1], size_input[2], 3])
-        im_flowline = np.zeros(size_output)
-        theta_offset = np.atleast_1d(theta_offset)
-    
-        if greyscale is True:
-            for a0 in range(size_input[0]):
-                if greyscale_max is True:
-                    im = np.max(orient_flowlines[a0, :, :, :], axis=2)
-                else:
-                    im = np.mean(orient_flowlines[a0, :, :, :], axis=2)
-    
-                sig = np.clip((im - int_range[0]) / (int_range[1] - int_range[0]), 0, 1)
-    
-                if power_scaling != 1:
-                    sig = sig**power_scaling
-    
-                if white_background is False:
-                    im_flowline[a0, :, :, :] = sig[:, :, None]
-                else:
-                    im_flowline[a0, :, :, :] = 1 - sig[:, :, None]
-    
-        else:
-            # Color basis
-            c0 = np.array([1.0, 0.0, 0.0])
-            c1 = np.array([0.0, 0.7, 0.0])
-            c2 = np.array([0.0, 0.3, 1.0])
-    
-            # angles
-            theta = np.linspace(0, np.pi, size_input[3], endpoint=False)
-            # Negate so the hue handedness matches the drawn flowline direction in the
-            # displayed (y-down) map: red horizontal, cyan vertical, "/" yellow, "\" purple.
-            theta_color = -theta * sym_rotation_order
-    
-            if size_input[0] > 1 and len(theta_offset) == 1:
-                theta_offset = np.ones(size_input[0]) * theta_offset
-    
-            for a0 in range(size_input[0]):
-                # color projections
-                b0 = np.maximum(
-                    1
-                    - np.abs(
-                        np.mod(theta_offset[a0] + theta_color + np.pi, 2 * np.pi) - np.pi
-                    )
-                    ** 2
-                    / (np.pi * 2 / 3) ** 2,
-                    0,
-                )
-                b1 = np.maximum(
-                    1
-                    - np.abs(
-                        np.mod(
-                            theta_offset[a0] + theta_color - np.pi * 2 / 3 + np.pi,
-                            2 * np.pi,
-                        )
-                        - np.pi
-                    )
-                    ** 2
-                    / (np.pi * 2 / 3) ** 2,
-                    0,
-                )
-                b2 = np.maximum(
-                    1
-                    - np.abs(
-                        np.mod(
-                            theta_offset[a0] + theta_color - np.pi * 4 / 3 + np.pi,
-                            2 * np.pi,
-                        )
-                        - np.pi
-                    )
-                    ** 2
-                    / (np.pi * 2 / 3) ** 2,
-                    0,
-                )
-    
-                sig = np.clip(
-                    (orient_flowlines[a0, :, :, :] - int_range[0])
-                    / (int_range[1] - int_range[0]),
-                    0,
-                    1,
-                )
-                if power_scaling != 1:
-                    sig = sig**power_scaling
-    
-                im_flowline[a0, :, :, :] = (
-                    np.sum(sig * b0[None, None, :], axis=2)[:, :, None] * c0[None, None, :]
-                    + np.sum(sig * b1[None, None, :], axis=2)[:, :, None]
-                    * c1[None, None, :]
-                    + np.sum(sig * b2[None, None, :], axis=2)[:, :, None]
-                    * c2[None, None, :]
-                )
-    
-                # clip limits
-                im_flowline[a0, :, :, :] = np.clip(im_flowline[a0, :, :, :], 0, 1)
-    
-                # contrast flip
-                if white_background is True:
-                    im = rgb_to_hsv(im_flowline[a0])
-                    im_v = im[:, :, 2]
-                    im[:, :, 1] = im_v
-                    im[:, :, 2] = 1
-                    im_flowline[a0] = hsv_to_rgb(im)
-    
-        if sum_radial_bins is True:
-            if white_background is False:
-                im_flowline = np.clip(np.sum(im_flowline, axis=0), 0, 1)[None, :, :, :]
-            else:
-                # im_flowline = np.clip(np.sum(im_flowline,axis=0)+1-im_flowline.shape[0],0,1)[None,:,:,:]
-                im_flowline = np.min(im_flowline, axis=0)[None, :, :, :]
-    
-        if plot_images is True:
-            if figsize is None:
-                fig, ax = plt.subplots(
-                    im_flowline.shape[0], 1, figsize=(10, im_flowline.shape[0] * 10)
-                )
-            else:
-                fig, ax = plt.subplots(im_flowline.shape[0], 1, figsize=figsize)
-    
-            if im_flowline.shape[0] > 1:
-                for a0 in range(im_flowline.shape[0]):
-                    ax[a0].imshow(im_flowline[a0])
-                    # ax[a0].axis('off')
-                plt.subplots_adjust(wspace=0, hspace=0.02)
-            else:
-                ax.imshow(im_flowline[0])
-                # ax.axis('off')
-            plt.show()
-    
-        return im_flowline
-    
-    
-    def make_flowline_rainbow_legend(
-        self,
-        im_size=np.array([256, 256]),
-        sym_rotation_order=2,
-        theta_offset_degrees=0.0,
-        white_background=False,
-        return_image=False,
-        radial_range=np.array([0.45, 0.9]),
-        plot_legend=True,
-        figsize=(4, 4),
-    ):
-        """
-        This function generates a legend for a the rainbow colored flowline maps, and returns it as an RGB image.
-    
-        Parameters
-        ----------
-        im_size (np.array):
-            Size of legend image in pixels.
-        sym_rotation_order (int):
-            rotational symmety for colouring
-        theta_offset_degrees (float):
-            Offset the anglular coloring by this value in degrees.
-            Rotation is Q with respect to R, in the positive (counter clockwise) direction.
-        white_background (bool):
-            For either color or greyscale output, switch to white background (from black).
-        return_image (bool):
-            Return the image array.
-        radial_range (np.array):
-            Inner and outer radius for the legend ring.
-        plot_legend (bool):
-            Plot the generated legend.
-        figsize (tuple or list):
-            Size of the plotted legend.
-    
-        Returns
-        ----------
-    
-        im_legend (array):
-            Image array for the legend.
-        """
-    
-        # Coordinates
-        x = np.linspace(-1, 1, im_size[0])
-        y = np.linspace(-1, 1, im_size[1])
-        ya, xa = np.meshgrid(y, x)
-        # TODO: Can replace with squared term? ra2? Faster
-        # ra = np.sqrt(xa**2 + ya**2)
-        ra2 = xa**2 + ya**2
-        ta = np.arctan2(ya, xa) + np.deg2rad(theta_offset_degrees)
-        ta_sym = ta * sym_rotation_order
-    
-        # mask
-        mask = np.logical_and(ra2 > radial_range[0]**2, ra2 < radial_range[1]**2)
-        # mask = np.logical_and(ra > radial_range[0], ra < radial_range[1])
-    
-        # rgb image
-        z = mask * np.exp(1j * ta_sym)
-        # hue_offset = 0
-        amp = np.abs(z)
-        vmin = np.min(amp)
-        vmax = np.max(amp)
-        ph = np.angle(z)  # + hue_offset
-        h = np.mod(ph / (2 * np.pi), 1)
-        s = 0.85 * np.ones_like(h)
-        v = (amp - vmin) / (vmax - vmin)
-        im_legend = hsv_to_rgb(np.dstack((h, s, v)))
-    
-        if white_background is True:
-            im_legend[im_legend.sum(2) == 0] = 1
-    
-        # plotting
-        if plot_legend:
-            fig, ax = plt.subplots(1, 1, figsize=figsize)
-            ax.imshow(im_legend)
-            ax.invert_yaxis()
-            # ax.set_axis_off()
-            ax.axis("off")
-    
-        if return_image:
-            return im_legend
-    
-    
-    def make_flowline_combined_image(
-        self,
-        orient_flowlines,
-        int_range=[0, 0.2],
-        cvals=np.array(
-            [
-                [0.0, 0.7, 0.0],
-                [1.0, 0.0, 0.0],
-                [0.0, 0.7, 1.0],
-            ]
-        ),
-        white_background=False,
-        power_scaling=1.0,
-        sum_radial_bins=True,
-        plot_images=True,
-        figsize=None,
-    ):
-        """
-        Generate RGB output images from the flowline arrays.
-    
-        Args:
-            orient_flowline (array):    Histogram of all orientations with coordinates [x y radial_bin theta]
-                                        We assume theta bin ranges from 0 to 180 degrees and is periodic.
-            int_range (float)           2 element array giving the intensity range
-            cvals (array):              Nx3 size array containing RGB colors for different radial ibns.
-            white_background (bool):    For either color or greyscale output, switch to white background (from black).
-            power_scaling (float):      Power law scaling for flowline intensities.
-            sum_radial_bins (bool):     Sum outputs over radial bins.
-            plot_images (bool):         Plot the output images for quick visualization.
-            figsize (2-tuple):          Size of output figure.
-    
-        Returns:
-            im_flowline (array):        flowline images
-        """
-    
-        # init array
-        size_input = orient_flowlines.shape
-        size_output = np.array([size_input[0], size_input[1], size_input[2], 3])
-        im_flowline = np.zeros(size_output)
-        cvals = np.array(cvals)
-    
-        # Generate all color images
-        for a0 in range(size_input[0]):
-            sig = np.clip(
-                (np.sum(orient_flowlines[a0, :, :, :], axis=2) - int_range[0])
-                / (int_range[1] - int_range[0]),
-                0,
-                1,
-            )
-            if power_scaling != 1:
-                sig = sig**power_scaling
-    
-            if white_background:
-                im_flowline[a0, :, :, :] = 1 - sig[:, :, None] * (
-                    1 - cvals[a0, :][None, None, :]
-                )
-            else:
-                im_flowline[a0, :, :, :] = sig[:, :, None] * cvals[a0, :][None, None, :]
-    
-            # # contrast flip
-            # if white_background is True:
-            #     im = rgb_to_hsv(im_flowline[a0,:,:,:])
-            #     # im_s = im[:,:,1]
-            #     im_v = im[:,:,2]
-            #     v_range = [np.min(im_v), np.max(im_v)]
-            #     print(v_range)
-    
-            #     im[:,:,1] = im_v
-            #     im[:,:,2] = 1
-            #     im_flowline[a0,:,:,:] = hsv_to_rgb(im)
-    
-        if sum_radial_bins is True:
-            if white_background is False:
-                im_flowline = np.clip(np.sum(im_flowline, axis=0), 0, 1)[None, :, :, :]
-            else:
-                # im_flowline = np.clip(np.sum(im_flowline,axis=0)+1-im_flowline.shape[0],0,1)[None,:,:,:]
-                im_flowline = np.min(im_flowline, axis=0)[None, :, :, :]
-    
-        if plot_images is True:
-            if figsize is None:
-                fig, ax = plt.subplots(
-                    im_flowline.shape[0], 1, figsize=(10, im_flowline.shape[0] * 10)
-                )
-            else:
-                fig, ax = plt.subplots(im_flowline.shape[0], 1, figsize=figsize)
-    
-            if im_flowline.shape[0] > 1:
-                for a0 in range(im_flowline.shape[0]):
-                    ax[a0].imshow(im_flowline[a0])
-                    ax[a0].axis("off")
-                plt.subplots_adjust(wspace=0, hspace=0.02)
-            else:
-                ax.imshow(im_flowline[0])
-                ax.axis("off")
-            plt.show()
-    
-        return im_flowline
+    def make_flowline_map(self, *args, **kwargs):
+        """Trace flowlines through an orientation histogram.
 
-    def get_intensity(
-        self, 
-        orient, 
-        x, 
-        y, 
-        t
-    ):
-        # utility function to get histogram intensites
-    
-        x = np.clip(x, 0, orient.shape[0] - 2)
-        y = np.clip(y, 0, orient.shape[1] - 2)
-    
-        xF = np.floor(x).astype("int")
-        yF = np.floor(y).astype("int")
-        tF = np.floor(t).astype("int")
-        dx = x - xF
-        dy = y - yF
-        dt = t - tF
-        t1 = np.mod(tF, orient.shape[2])
-        t2 = np.mod(tF + 1, orient.shape[2])
-    
-        int_vals = (
-            orient[xF, yF, t1] * ((1 - dx) * (1 - dy) * (1 - dt))
-            + orient[xF, yF, t2] * ((1 - dx) * (1 - dy) * (dt))
-            + orient[xF, yF + 1, t1] * ((1 - dx) * (dy) * (1 - dt))
-            + orient[xF, yF + 1, t2] * ((1 - dx) * (dy) * (dt))
-            + orient[xF + 1, yF, t1] * ((dx) * (1 - dy) * (1 - dt))
-            + orient[xF + 1, yF, t2] * ((dx) * (1 - dy) * (dt))
-            + orient[xF + 1, yF + 1, t1] * ((dx) * (dy) * (1 - dt))
-            + orient[xF + 1, yF + 1, t2] * ((dx) * (dy) * (dt))
-        )
-    
-        return int_vals
-    
-    
-    def set_intensity(
-        self, 
-        orient, 
-        xy_t_int
-    ):
-        # utility function to set flowline intensites
-    
-        xF = np.floor(xy_t_int[:, 0]).astype("int")
-        yF = np.floor(xy_t_int[:, 1]).astype("int")
-        tF = np.floor(xy_t_int[:, 2]).astype("int")
-        dx = xy_t_int[:, 0] - xF
-        dy = xy_t_int[:, 1] - yF
-        dt = xy_t_int[:, 2] - tF
-    
-        inds_1D = np.ravel_multi_index(
-            [xF, yF, tF], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (1 - dx) * (
-            1 - dy
-        ) * (1 - dt)
-        inds_1D = np.ravel_multi_index(
-            [xF, yF, tF + 1], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (1 - dx) * (
-            1 - dy
-        ) * (dt)
-        inds_1D = np.ravel_multi_index(
-            [xF, yF + 1, tF], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (1 - dx) * (
-            dy
-        ) * (1 - dt)
-        inds_1D = np.ravel_multi_index(
-            [xF, yF + 1, tF + 1], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (1 - dx) * (
-            dy
-        ) * (dt)
-        inds_1D = np.ravel_multi_index(
-            [xF + 1, yF, tF], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (dx) * (
-            1 - dy
-        ) * (1 - dt)
-        inds_1D = np.ravel_multi_index(
-            [xF + 1, yF, tF + 1], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (dx) * (
-            1 - dy
-        ) * (dt)
-        inds_1D = np.ravel_multi_index(
-            [xF + 1, yF + 1, tF], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (dx) * (dy) * (
-            1 - dt
-        )
-        inds_1D = np.ravel_multi_index(
-            [xF + 1, yF + 1, tF + 1], orient.shape[0:3], mode=["clip", "clip", "wrap"]
-        )
-        orient.ravel()[inds_1D] = orient.ravel()[inds_1D] + xy_t_int[:, 3] * (dx) * (dy) * (
-            dt
-        )
-    
-        return orient
+        Forwards to :func:`quantem.diffraction.orientation.make_flowline_map`.
+        """
+
+        return _make_flowline_map(*args, **kwargs)
+
+    def make_flowline_rainbow_image(self, *args, **kwargs):
+        """Render flowlines as an orientation-coloured image.
+
+        Forwards to :func:`quantem.diffraction.orientation.make_flowline_rainbow_image`.
+        """
+
+        return _make_flowline_rainbow_image(*args, **kwargs)
+
+    def make_flowline_rainbow_legend(self, *args, **kwargs):
+        """Colour key for the rainbow flowline image.
+
+        Forwards to :func:`quantem.diffraction.orientation.make_flowline_rainbow_legend`.
+        """
+
+        return _make_flowline_rainbow_legend(*args, **kwargs)
+
+    def make_flowline_combined_image(self, *args, **kwargs):
+        """Render flowlines with explicit per-bin colours.
+
+        Forwards to :func:`quantem.diffraction.orientation.make_flowline_combined_image`.
+        """
+
+        return _make_flowline_combined_image(*args, **kwargs)
+
+    def get_intensity(self, *args, **kwargs):
+        """Sample an orientation histogram with linear interpolation.
+
+        Forwards to :func:`quantem.diffraction.orientation._get_intensity`.
+        """
+
+        return _get_intensity(*args, **kwargs)
+
+    def set_intensity(self, *args, **kwargs):
+        """Deposit interpolated weight into an orientation histogram.
+
+        Forwards to :func:`quantem.diffraction.orientation._set_intensity`.
+        """
+
+        return _set_intensity(*args, **kwargs)
 
     def interactive_probe_selector(self, probe_map=None, figsize=(14, 8), cmap='viridis'):
         """
