+++ src/quantem/diffraction/orientation_correlation.py
@@ -4,7 +4,9 @@ from __future__ import annotations
 
 from collections.abc import Sequence
 
+import matplotlib.pyplot as plt
 import numpy as np
+from scipy.ndimage import gaussian_filter1d
 import torch
 from scipy import fft as scipy_fft
 from tqdm.auto import tqdm
@@ -592,3 +594,458 @@ def calculate_orientation_correlation(
             output = output.detach().cpu().numpy()
 
     return output, pair_indices
+
+
+# Default 1/e decay length for the orientation-correlation slope fit, as a
+# fraction of the fitted lobe's distance span. Chosen on real scan data
+# (pg3T2 07-03-2024/61): it holds the fitted intercept within ~1.7 degrees of
+# the measured boundary at zero separation across all six ring pairs, versus
+# up to 7 degrees for an unweighted fit, while keeping enough effective points
+# for a stable slope. See plot_orientation_correlation(slope_weight_scale=...).
+SLOPE_WEIGHT_FRACTION = 0.10
+
+
+def plot_orientation_correlation(
+    orient_corr,
+    orient_corr_pairs=None,
+    *,
+    pair_indices=None,
+    pixel_size=1.0,
+    pixel_units="scan pixels",
+    probability_range=(0.5, 2.0),
+    cmap="correlation",
+    figsize=None,
+    show_metrics=True,
+    return_metrics=False,
+    slope_weight_scale=None,
+):
+    """Plot distance-orientation correlations using Matplotlib.
+
+    The 50% boundary is halfway between the correlation at zero separation
+    and the random-association baseline of one. Its intercepts give the
+    radial and annular 50% distances. The signed slope is fitted separately
+    to the primary correlation-equals-one boundary between positive
+    correlation and anticorrelation.
+
+    Parameters
+    ----------
+    slope_weight_scale : float, optional
+        1/e decay length, in ``pixel_units``, of the exponential weighting
+        applied to the slope fit. The correlation-equals-one boundary
+        saturates with distance, so an unweighted straight line over the
+        whole lobe is dominated by the flat tail: it biases the slope low
+        and pushes the fitted intercept off the measured boundary at zero
+        separation. Weighting towards short distances makes ``slope`` the
+        near-origin tangent instead. Defaults to
+        ``SLOPE_WEIGHT_FRACTION`` times the fitted distance span. Pass
+        ``numpy.inf`` to restore the previous unweighted full-lobe fit.
+
+    Notes
+    -----
+    Each metrics entry reports ``slope_fit_intercept_degrees`` (compare it
+    against the boundary at zero separation to check the fit),
+    ``slope_fit_effective_point_count`` (Kish effective sample size, which
+    falls as the weighting sharpens), and the resolved
+    ``slope_weight_scale``. ``slope_fit_r_squared`` is weighted with the
+    same weights, so it is not comparable to an unweighted R-squared.
+    """
+    from matplotlib.colors import LinearSegmentedColormap, LogNorm
+    from matplotlib.lines import Line2D
+
+    def crossing(coordinates, profile, level):
+        profile = np.asarray(profile, dtype=float)
+        coordinates = np.asarray(coordinates, dtype=float)
+        if not np.isfinite(profile[0]):
+            return np.nan
+        initial_side = profile[0] - level
+        if initial_side == 0:
+            return float(coordinates[0])
+        for point in range(1, len(profile)):
+            before, after = profile[point - 1], profile[point]
+            if not np.isfinite(before) or not np.isfinite(after):
+                continue
+            before_side = before - level
+            after_side = after - level
+            if before_side == 0:
+                return float(coordinates[point - 1])
+            if before_side * after_side <= 0:
+                if before == after:
+                    return float(coordinates[point])
+                fraction = -before_side / (after_side - before_side)
+                return float(
+                    coordinates[point - 1]
+                    + fraction * (coordinates[point] - coordinates[point - 1])
+                )
+        return np.nan
+
+    values = orient_corr
+    if values is None:
+        raise RuntimeError(
+            "No orientation correlation is available. Run "
+            "calculate_orientation_correlation() first or pass orient_corr."
+        )
+    values = np.asarray(values)
+    if values.ndim != 3:
+        raise ValueError(
+            "orient_corr must have shape (pair, relative_angle, distance)."
+        )
+    if values.shape[0] == 0:
+        raise ValueError("orient_corr must contain at least one radial-bin pair.")
+    if values.shape[1] < 2 or values.shape[2] < 2:
+        raise ValueError(
+            "orient_corr requires at least two angle and two distance samples."
+        )
+    labels = orient_corr_pairs if pair_indices is None else pair_indices
+    if labels is not None:
+        labels = np.asarray(labels)
+        if labels.shape != (values.shape[0], 2):
+            raise ValueError(
+                f"pair_indices must have shape ({values.shape[0]}, 2)."
+            )
+
+    panel_count = values.shape[0]
+    column_count = min(3, max(1, panel_count))
+    row_count = int(np.ceil(panel_count / column_count))
+    if figsize is None:
+        figsize = (4.5 * column_count, 3.8 * row_count)
+    fig, axes = plt.subplots(
+        row_count,
+        column_count,
+        figsize=figsize,
+        squeeze=False,
+        constrained_layout=True,
+    )
+    lower, upper = map(float, probability_range)
+    if not 0 < lower < upper:
+        raise ValueError("probability_range must satisfy 0 < lower < upper.")
+    if cmap == "correlation":
+        cmap = LinearSegmentedColormap.from_list(
+            "quantem_correlation",
+            [
+                (0.00, "#002b9a"),
+                (0.32, "#1769e8"),
+                (0.50, "#b8b8b8"),
+                (0.68, "#f23838"),
+                (1.00, "#9e0015"),
+            ],
+        )
+    distance_max = (values.shape[2] - 1) * float(pixel_size)
+    distances = np.arange(values.shape[2], dtype=float) * float(pixel_size)
+    angles = np.linspace(0.0, 180.0, values.shape[1])
+    image = None
+    metrics = []
+    for index, ax in enumerate(axes.flat):
+        if index >= panel_count:
+            ax.set_visible(False)
+            continue
+        image = ax.imshow(
+            values[index],
+            origin="lower",
+            aspect="auto",
+            extent=(0, distance_max, 0, 180),
+            norm=LogNorm(vmin=lower, vmax=upper),
+            cmap=cmap,
+        )
+        if labels is None:
+            title = f"Ring pair {index}"
+            pair = (index, index)
+        else:
+            pair = tuple(int(value) for value in labels[index])
+            title = (
+                f"Autocorrelation of Ring {pair[0]}"
+                if pair[0] == pair[1]
+                else f"Correlation of Rings {pair[0]} and {pair[1]}"
+            )
+        ax.set(
+            title=title,
+            xlabel=f"distance ({pixel_units})",
+            ylabel="relative orientation (degrees)",
+        )
+        panel = np.asarray(values[index], dtype=float)
+        origin_probability = panel[0, 0]
+        half_probability = (
+            1.0 + 0.5 * (origin_probability - 1.0)
+            if np.isfinite(origin_probability)
+            and origin_probability > 0.0
+            and not np.isclose(origin_probability, 1.0)
+            else np.nan
+        )
+        radial_distance = np.nan
+        annular_distance = np.nan
+        slope = np.nan
+        slope_fit_r_squared = np.nan
+        slope_fit_point_count = 0
+        slope_fit_intercept = np.nan
+        slope_fit_effective_count = np.nan
+        weight_scale = np.nan
+        fit_distances = np.array([])
+        fit_angles = np.array([])
+        if np.isfinite(half_probability):
+            radial_distance = crossing(
+                distances, panel[0, :], half_probability
+            )
+            annular_distance = crossing(
+                angles, panel[:, 0], half_probability
+            )
+            if np.isfinite(radial_distance):
+                ax.scatter(
+                    [radial_distance],
+                    [0],
+                    marker="o",
+                    s=45,
+                    facecolor="white",
+                    edgecolor="black",
+                    linewidth=0.8,
+                    zorder=5,
+                )
+            if np.isfinite(annular_distance):
+                ax.scatter(
+                    [0],
+                    [annular_distance],
+                    marker="D",
+                    s=40,
+                    facecolor="white",
+                    edgecolor="black",
+                    linewidth=0.8,
+                    zorder=5,
+                )
+
+        # The slope belongs to the gray probability/random = 1 boundary, not
+        # the half-maximum contour used for the two distance intercepts.
+        baseline_boundary = np.array(
+            [
+                crossing(angles, panel[:, radius], 1.0)
+                for radius in range(panel.shape[1])
+            ]
+        )
+        baseline_radial_intercept = crossing(distances, panel[0, :], 1.0)
+        valid = np.isfinite(baseline_boundary)
+        if np.isfinite(baseline_radial_intercept):
+            valid &= distances <= baseline_radial_intercept + float(pixel_size)
+
+        # Select the earliest contiguous run: this follows the principal
+        # red/blue lobe from the angular axis and rejects remote closed loops.
+        valid_indices = np.flatnonzero(valid)
+        primary_indices = np.array([], dtype=int)
+        if valid_indices.size:
+            angular_jump = np.abs(
+                np.diff(baseline_boundary[valid_indices])
+            )
+            maximum_step = max(15.0, 4.0 * (angles[1] - angles[0]))
+            split_points = np.flatnonzero(
+                (np.diff(valid_indices) > 1) | (angular_jump > maximum_step)
+            ) + 1
+            segments = np.split(
+                valid_indices, split_points
+            )
+            primary_indices = next(
+                (segment for segment in segments if len(segment) >= 2),
+                np.array([], dtype=int),
+            )
+        if primary_indices.size >= 5:
+            # A connected correlation=1 contour can rise away from the
+            # origin, turn around at large distance, and return as part of
+            # the same loop. Fitting that entire loop can reverse the sign
+            # of the visually obvious near-origin boundary. Stop at the
+            # first sustained turning point while tolerating isolated
+            # pixel-scale contour noise.
+            boundary_run = baseline_boundary[primary_indices]
+            smoothing_sigma = min(3.0, max(0.75, len(boundary_run) / 100.0))
+            boundary_smooth = gaussian_filter1d(
+                boundary_run, sigma=smoothing_sigma, mode="nearest"
+            )
+            boundary_gradient = np.gradient(boundary_smooth)
+            initial_count = min(20, max(3, len(boundary_run) // 10))
+            initial_trend = float(
+                np.median(boundary_gradient[:initial_count])
+            )
+            if not np.isclose(initial_trend, 0.0):
+                reversal = boundary_gradient * np.sign(initial_trend) < 0
+                persistence = min(5, max(2, len(boundary_run) // 20))
+                sustained = np.convolve(
+                    reversal.astype(int),
+                    np.ones(persistence, dtype=int),
+                    mode="valid",
+                )
+                turning_points = np.flatnonzero(sustained == persistence)
+                if turning_points.size and turning_points[0] >= 2:
+                    primary_indices = primary_indices[
+                        : turning_points[0] + 1
+                    ]
+        if primary_indices.size:
+            fit_distances = distances[primary_indices]
+            fit_values = baseline_boundary[primary_indices]
+
+            # The correlation=1 boundary saturates: it climbs steeply near the
+            # origin and flattens at large separation. An unweighted straight
+            # line over the whole lobe is therefore dominated by the flat tail,
+            # which drags the intercept off the measured boundary at d=0 (up to
+            # ~7 degrees on real data, visibly landing in the blue region) and
+            # biases the slope low. Weight the fit towards short distances so
+            # `slope` is the near-origin tangent, which is the physically
+            # meaningful quantity.
+            fit_span = float(fit_distances.max() - fit_distances.min())
+            if slope_weight_scale is None:
+                weight_scale = SLOPE_WEIGHT_FRACTION * fit_span
+            else:
+                weight_scale = float(slope_weight_scale)
+            if not np.isfinite(weight_scale) or weight_scale <= 0:
+                # np.inf (or a non-positive scale) restores the legacy
+                # unweighted fit over the full lobe.
+                fit_weights = np.ones_like(fit_distances)
+                weight_scale = np.inf
+            else:
+                fit_weights = np.exp(
+                    -(fit_distances - fit_distances.min()) / weight_scale
+                )
+
+            root_weights = np.sqrt(fit_weights)
+            design = (
+                np.vstack([fit_distances, np.ones_like(fit_distances)]).T
+                * root_weights[:, None]
+            )
+            fit_slope, fit_intercept = np.linalg.lstsq(
+                design, fit_values * root_weights, rcond=None
+            )[0]
+            fit_angles = fit_intercept + fit_slope * fit_distances
+            slope = float(fit_slope)
+            slope_fit_intercept = float(fit_intercept)
+            slope_fit_point_count = int(fit_distances.size)
+            # Kish effective sample size: how many points the weighting really
+            # uses, so a too-aggressive scale is visible rather than silent.
+            slope_fit_effective_count = float(
+                fit_weights.sum() ** 2 / np.sum(fit_weights**2)
+            )
+            weight_mean = float(
+                np.average(fit_values, weights=fit_weights)
+            )
+            residual_sum_squares = float(
+                np.sum(fit_weights * (fit_values - fit_angles) ** 2)
+            )
+            total_sum_squares = float(
+                np.sum(fit_weights * (fit_values - weight_mean) ** 2)
+            )
+            slope_fit_r_squared = (
+                1.0 - residual_sum_squares / total_sum_squares
+                if total_sum_squares > 0
+                else np.nan
+            )
+
+        if fit_distances.size:
+            # Draw only where the weighting actually constrains the line. Past
+            # ~3 decay lengths the boundary has saturated away from this
+            # tangent, and extending the line there would misrepresent the fit.
+            if np.isfinite(weight_scale):
+                drawn = fit_distances <= (
+                    fit_distances.min() + 3.0 * weight_scale
+                )
+                if drawn.sum() < 2:
+                    drawn = np.zeros_like(fit_distances, dtype=bool)
+                    drawn[: min(2, drawn.size)] = True
+            else:
+                drawn = np.ones_like(fit_distances, dtype=bool)
+            visible_fit = drawn & (fit_angles >= 0) & (fit_angles <= 180)
+            ax.plot(
+                fit_distances[visible_fit],
+                fit_angles[visible_fit],
+                color="#ffe600",
+                linestyle="-",
+                linewidth=2.5,
+                zorder=6,
+            )
+        ax.legend(
+            handles=[
+                Line2D(
+                    [0],
+                    [0],
+                    marker="o",
+                    color="none",
+                    markerfacecolor="white",
+                    markeredgecolor="black",
+                    label="50% radial intercept",
+                ),
+                Line2D(
+                    [0],
+                    [0],
+                    marker="D",
+                    color="none",
+                    markerfacecolor="white",
+                    markeredgecolor="black",
+                    label="50% annular intercept",
+                ),
+                Line2D(
+                    [0],
+                    [0],
+                    color="#ffe600",
+                    linewidth=2.5,
+                    label="near-origin baseline fit",
+                ),
+            ],
+            loc="lower right",
+            fontsize=7,
+            framealpha=0.82,
+        )
+
+        panel_metrics = {
+            "pair": pair,
+            "title": title,
+            "half_probability": float(half_probability),
+            "radial_distance": float(radial_distance),
+            "annular_distance_degrees": float(annular_distance),
+            "slope_degrees_per_unit": float(slope),
+            "slope_fit_r_squared": float(slope_fit_r_squared),
+            "slope_fit_point_count": slope_fit_point_count,
+            "slope_fit_intercept_degrees": float(slope_fit_intercept),
+            "slope_fit_effective_point_count": float(
+                slope_fit_effective_count
+            ),
+            "slope_weight_scale": float(weight_scale),
+            "slope_contour_probability": 1.0,
+            "distance_units": pixel_units,
+        }
+        metrics.append(panel_metrics)
+        if show_metrics:
+            radial_text = (
+                f"{radial_distance:.2f} {pixel_units}"
+                if np.isfinite(radial_distance)
+                else "not resolved"
+            )
+            annular_text = (
+                f"{annular_distance:.2f} degrees"
+                if np.isfinite(annular_distance)
+                else "not resolved"
+            )
+            slope_text = (
+                f"{slope:.2f} degrees/{pixel_units}"
+                if np.isfinite(slope)
+                else "not resolved"
+            )
+            ax.text(
+                0.98,
+                0.98,
+                "50% radial distance = "
+                + radial_text
+                + "\n50% annular distance = "
+                + annular_text
+                + "\nslope = "
+                + slope_text,
+                transform=ax.transAxes,
+                ha="right",
+                va="top",
+                fontsize=8,
+                bbox={
+                    "boxstyle": "round,pad=0.35",
+                    "facecolor": "white",
+                    "edgecolor": "black",
+                    "alpha": 0.82,
+                },
+            )
+    if image is not None:
+        fig.colorbar(
+            image,
+            ax=[ax for ax in axes.flat if ax.get_visible()],
+            label="probability / random",
+        )
+    if return_metrics:
+        return fig, axes, metrics
+    return fig, axes
